# executor: parallelize aggregations with distinct in hashAgg