--- a/executor/aggfuncs/func_count_distinct.go
+++ b/executor/aggfuncs/func_count_distinct.go
@@ -324,6 +324,90 @@ func (e *countOriginalWithDistinct) UpdatePartialResult(sctx sessionctx.Context,
 	return 0, nil
 }
 
+type countPartialWithDistinct4Int struct {
+	countOriginalWithDistinct4Int
+}
+
+func (e *countPartialWithDistinct4Int) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4CountDistinctInt)(src), (*partialResult4CountDistinctInt)(dst)
+	for k := range p1.valSet {
+		if !p2.valSet.Exist(k) {
+			p2.valSet.Insert(k)
+		}
+	}
+	return 0, nil
+}
+
+type countPartialWithDistinct4Real struct {
+	countOriginalWithDistinct4Real
+}
+
+func (e *countPartialWithDistinct4Real) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4CountDistinctReal)(src), (*partialResult4CountDistinctReal)(dst)
+	for k := range p1.valSet {
+		if !p2.valSet.Exist(k) {
+			p2.valSet.Insert(k)
+		}
+	}
+	return 0, nil
+}
+
+type countPartialWithDistinct4Decimal struct {
+	countOriginalWithDistinct4Decimal
+}
+
+func (e *countPartialWithDistinct4Decimal) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4CountDistinctDecimal)(src), (*partialResult4CountDistinctDecimal)(dst)
+	for k := range p1.valSet {
+		if !p2.valSet.Exist(k) {
+			p2.valSet.Insert(k)
+		}
+	}
+	return 0, nil
+}
+
+type countPartialWithDistinct4Duration struct {
+	countOriginalWithDistinct4Duration
+}
+
+func (e *countPartialWithDistinct4Duration) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4CountDistinctDuration)(src), (*partialResult4CountDistinctDuration)(dst)
+	for k := range p1.valSet {
+		if !p2.valSet.Exist(k) {
+			p2.valSet.Insert(k)
+		}
+	}
+	return 0, nil
+}
+
+type countPartialWithDistinct4String struct {
+	countOriginalWithDistinct4String
+}
+
+func (e *countPartialWithDistinct4String) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4CountDistinctString)(src), (*partialResult4CountDistinctString)(dst)
+	for k := range p1.valSet {
+		if !p2.valSet.Exist(k) {
+			p2.valSet.Insert(k)
+		}
+	}
+	return 0, nil
+}
+
+type countPartialWithDistinct struct {
+	countOriginalWithDistinct
+}
+
+func (e *countPartialWithDistinct) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4CountWithDistinct)(src), (*partialResult4CountWithDistinct)(dst)
+	for k := range p1.valSet {
+		if !p2.valSet.Exist(k) {
+			p2.valSet.Insert(k)
+		}
+	}
+	return 0, nil
+}
+
 // evalAndEncode eval one row with an expression and encode value to bytes.
 func evalAndEncode(
 	sctx sessionctx.Context, arg expression.Expression,