--- a/executor/aggfuncs/func_group_concat.go
+++ b/executor/aggfuncs/func_group_concat.go
@@ -154,6 +154,8 @@ type partialResult4GroupConcatDistinct struct {
 	basePartialResult4GroupConcat
 	valSet            set.StringSet
 	encodeBytesBuffer []byte
+	valList           []string
+	joinedValList     []string
 }
 
 type groupConcatDistinct struct {
@@ -164,12 +166,14 @@ func (e *groupConcatDistinct) AllocPartialResult() (pr PartialResult, memDelta i
 	p := new(partialResult4GroupConcatDistinct)
 	p.valsBuf = &bytes.Buffer{}
 	p.valSet = set.NewStringSet()
+	p.valList, p.joinedValList = []string{}, []string{}
 	return PartialResult(p), 0
 }
 
 func (e *groupConcatDistinct) ResetPartialResult(pr PartialResult) {
 	p := (*partialResult4GroupConcatDistinct)(pr)
 	p.buffer, p.valSet = nil, set.NewStringSet()
+	p.valList, p.joinedValList = p.valList[:0], p.joinedValList[:0]
 }
 
 func (e *groupConcatDistinct) UpdatePartialResult(sctx sessionctx.Context, rowsInGroup []chunk.Row, pr PartialResult) (memDelta int64, err error) {
@@ -197,6 +201,8 @@ func (e *groupConcatDistinct) UpdatePartialResult(sctx sessionctx.Context, rowsI
 			continue
 		}
 		p.valSet.Insert(joinedVal)
+		p.valList = append(p.valList, p.valsBuf.String())
+		p.joinedValList = append(p.joinedValList, joinedVal)
 		// write separator
 		if p.buffer == nil {
 			p.buffer = &bytes.Buffer{}
@@ -208,6 +214,30 @@ func (e *groupConcatDistinct) UpdatePartialResult(sctx sessionctx.Context, rowsI
 	}
 	if p.buffer != nil {
 		return 0, e.truncatePartialResultIfNeed(sctx, p.buffer)
+	}
+	return 0, nil
+}
+
+func (e *groupConcatDistinct) MergePartialResult(sctx sessionctx.Context, src, dst PartialResult) (memDelta int64, err error) {
+	p1, p2 := (*partialResult4GroupConcatDistinct)(src), (*partialResult4GroupConcatDistinct)(dst)
+	for i := range p1.valList {
+		if p2.valSet.Exist(p1.joinedValList[i]) {
+			continue
+		}
+		p2.valSet.Insert(p1.joinedValList[i])
+		p2.valList = append(p2.valList, p1.valList[i])
+		p2.joinedValList = append(p2.joinedValList, p1.joinedValList[i])
+		// write separator
+		if p2.buffer == nil {
+			p2.buffer = &bytes.Buffer{}
+		} else {
+			p2.buffer.WriteString(e.sep)
+		}
+		// write values
+		p2.buffer.WriteString(p1.valList[i])
+	}
+	if p2.buffer != nil {
+		return 0, e.truncatePartialResultIfNeed(sctx, p2.buffer)
 	}
 	return 0, nil
 }