<!-- Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?

Issue Number: close #17075 <!-- REMOVE this line if no issue to close -->

Problem Summary:

``` sql
select count(distinct a) from t;
```

If one of the aggregate functions in a query needs distinct evaluation, the hashAgg executor will run in single thread.

### What is changed and how it works?

What's Changed:

- HashAgg executor will execute aggregation functions with distinct (without order by) in parallel instead of in single thread.

How it Works:

- modify the structures of `partialResult4XXXDistinct`s to store ordered distinct values
- add aggregation functions with distinct and methods `MergePartialResult` for final workers in order to merge the results of partial workers correctly.

### Related changes

- Need to cherry-pick to the release branch 3.0, 4.0

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test
- Integration test

Side effects

- Performance regression
    - Consumes more CPU
    - Consumes more MEM

Benchmark for `AggDistinct`:

```
name                                                                                                      old time/op    new time/op    delta
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:4)-12        3.58ms ± 7%    3.69ms ± 6%      ~     (p=0.105 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:100000,_concurrency:4)-12         40.9ms ± 2%    13.4ms ± 3%   -67.36%  (p=0.000 n=9+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:4)-12      1.50ms ± 2%    1.48ms ± 4%      ~     (p=0.089 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:100000,_concurrency:4)-12       12.5ms ± 2%    12.6ms ± 1%    +1.08%  (p=0.009 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:4)-12       24.5ms ± 3%    25.7ms ± 4%    +4.91%  (p=0.000 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:1000000,_concurrency:4)-12         477ms ± 3%     129ms ± 3%   -72.97%  (p=0.000 n=9+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:4)-12     14.1ms ± 2%    14.3ms ± 2%    +1.44%  (p=0.023 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:1000000,_concurrency:4)-12       134ms ± 2%     137ms ± 2%    +1.98%  (p=0.003 n=10+9)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-12       236ms ± 5%     248ms ± 7%    +5.15%  (p=0.005 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:10000000,_concurrency:4)-12        5.70s ± 2%     1.47s ± 6%   -74.20%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-12     100ms ± 4%     102ms ± 5%      ~     (p=0.052 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:10000000,_concurrency:4)-12      1.18s ± 3%     1.20s ± 3%      ~     (p=0.315 n=10+9)

name                                                                                                      old alloc/op   new alloc/op   delta
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:4)-12        4.63MB ± 1%    4.64MB ± 0%    +0.28%  (p=0.027 n=10+8)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:100000,_concurrency:4)-12         7.88MB ± 1%   16.98MB ± 0%  +115.44%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:4)-12      45.7kB ± 0%    45.7kB ± 0%      ~     (p=0.793 n=10+9)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:100000,_concurrency:4)-12       4.22MB ± 1%    4.24MB ± 1%      ~     (p=0.143 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:4)-12       26.3MB ± 0%    26.3MB ± 0%      ~     (p=0.353 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:1000000,_concurrency:4)-12        80.4MB ± 0%   165.7MB ± 0%  +106.00%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:4)-12     70.1kB ± 0%    70.1kB ± 0%      ~     (p=0.962 n=9+9)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:1000000,_concurrency:4)-12      48.0MB ± 0%    48.0MB ± 0%    +0.06%  (p=0.000 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-12       243MB ± 0%     243MB ± 0%      ~     (p=0.218 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:10000000,_concurrency:4)-12        710MB ± 0%    1758MB ± 0%  +147.60%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-12    70.0kB ± 0%    70.0kB ± 0%      ~     (p=0.877 n=9+8)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:10000000,_concurrency:4)-12      390MB ± 0%     390MB ± 0%    +0.10%  (p=0.000 n=10+10)

name                                                                                                      old allocs/op  new allocs/op  delta
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:4)-12         29.1k ± 0%     29.1k ± 0%    +0.01%  (p=0.048 n=10+9)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:100000,_concurrency:4)-12           223k ± 0%      110k ± 0%   -50.95%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:100000,_concurrency:4)-12        50.0 ± 0%      50.0 ± 0%      ~     (all equal)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:100000,_concurrency:4)-12        19.4k ± 0%     19.4k ± 0%    +0.10%  (p=0.008 n=9+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:4)-12        31.7k ± 0%     31.7k ± 0%      ~     (p=0.137 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:1000000,_concurrency:4)-12         2.07M ± 0%     0.27M ± 0%   -87.10%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:1000000,_concurrency:4)-12       51.0 ± 0%      51.0 ± 0%      ~     (all equal)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:1000000,_concurrency:4)-12       67.8k ± 0%     67.8k ± 0%      ~     (p=0.105 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-12       58.1k ± 0%     58.1k ± 0%      ~     (p=0.696 n=10+10)
AggDistinct/(execType:hash,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:10000000,_concurrency:4)-12        20.3M ± 0%      0.9M ± 0%   -95.74%  (p=0.000 n=10+10)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:false,_rows:10000000,_concurrency:4)-12      50.0 ± 0%      50.0 ± 0%      ~     (all equal)
AggDistinct/(execType:stream,_aggFunc:sum,_ndv:1000,_hasDistinct:true,_rows:10000000,_concurrency:4)-12       258k ± 0%      257k ± 0%      ~     (p=0.222 n=9+9)
```

### Release note <!-- bugfixes or new feature need a release note -->

- Parallelize evaluation for aggregate functions with distinct in hash aggregation
<!-- Please write a release note here to describe the change you made when it is released to the users of TiDB. If your PR doesn't involve any change to TiDB(like test enhancements, RFC proposals...), you can write `No release note`. -->