@zz-jason 2020-09-02
for performance improvements, it's better to do some benchmark and post the result in the PR description.